+++ ElementX/Sources/Screens/InviteUsers/InviteUsersModels.swift
@@ -23,7 +23,7 @@ enum InviteUsersViewModelAction {
 struct InviteUsersViewState: BindableState {
     var bindings = InviteUsersViewStateBindings()
     
-    var usersSection: SearchUsersUsersSection = .init(type: .suggestions, users: [])
+    var usersSection: SearchUsersSection = .init(type: .suggestions, users: [])
     var selectedUsers: [UserProfile] = []
     
     var isSearching: Bool {
